+++ src/main/java/org/sciborgs1155/robot/turret/SimTurret.java
@@ -41,8 +41,8 @@ public class SimTurret implements TurretIO {
   private final EasyCRTConfig crtConfig =
       new EasyCRTConfig(() -> Rotations.of(encoderA()), () -> Rotations.of(encoderB()))
           .withEncoderRatios(
-              (double) TURRET_GEARING / ENCODER_A_GEARING,
-              (double) TURRET_GEARING / ENCODER_B_GEARING)
+              TURRET_GEARING / ENCODER_A_GEARING,
+              TURRET_GEARING / ENCODER_B_GEARING)
           .withMechanismRange(MIN_ANGLE, MAX_ANGLE)
           .withMatchTolerance(CRT_MATCH_TOLERANCE);
 
@@ -56,7 +56,7 @@ public class SimTurret implements TurretIO {
   @Override
   public double encoderA() {
     double turretRot = trueAngleRad() / (2.0 * Math.PI);
-    double encoderRot = turretRot * ((double) TURRET_GEARING / ENCODER_A_GEARING);
+    double encoderRot = turretRot * (TURRET_GEARING / ENCODER_A_GEARING);
 
     return MathUtil.inputModulus(encoderRot, 0.0, 1.0);
   }
@@ -64,7 +64,7 @@ public class SimTurret implements TurretIO {
   @Override
   public double encoderB() {
     double turretRot = trueAngleRad() / (2.0 * Math.PI);
-    double encoderRot = turretRot * ((double) TURRET_GEARING / ENCODER_B_GEARING);
+    double encoderRot = turretRot * (TURRET_GEARING / ENCODER_B_GEARING);
 
     return MathUtil.inputModulus(encoderRot, 0.0, 1.0);
   }
